Fix pkgname fetching step
sed with a regex was used before, which just parsed the `pkgname=`
line and fetched the value. However, in some PKGBUILDs the `pkgname`
itself is formed from other variables, which happens to be the case in
this one.

To resolve it, we'd need to parse it and resolve the variable
references with more sed calls. Or, we can just source the PKGBUILD
which thankfully doesn't execute anything by itself, and from there
get the resolved by the shell `pkgname` variable.